[New] Added test cases for transitions
