+++ app/javascript/components/transitions/TransitionPanel.spec.js
@@ -1,0 +1,15 @@
+import { mountedComponent, screen } from "../../test-utils";
+
+import TransitionPanel from "./TransitionPanel";
+
+describe("<TransitionPanel />", () => {
+  const props = {
+    children: <p>This is a children</p>
+  };
+
+  it("renders a <p> as children of  TransitionPanel", () => {
+    mountedComponent(<TransitionPanel {...props} />);
+
+    expect(screen.getByText(/This is a children/i)).toBeInTheDocument();
+  });
+});
