--- app/javascript/components/transitions/TransitionStatus.unit.test.js
@@ -1,22 +1,0 @@
-// Copyright (c) 2014 - 2023 UNICEF. All rights reserved.
-
-import { Chip } from "@material-ui/core";
-
-import { setupMountedComponent } from "../../test";
-
-import TransitionStatus from "./TransitionStatus";
-
-describe("<TransitionStatus />", () => {
-  let component;
-  const props = {
-    status: "inprogress"
-  };
-
-  beforeEach(() => {
-    ({ component } = setupMountedComponent(TransitionStatus, props));
-  });
-
-  it("renders a Chip TransitionStatus", () => {
-    expect(component.find(TransitionStatus).find(Chip)).to.have.length(1);
-  });
-});
